Update.
	* sysdeps/unix/sysv/linux/ia64/lowlevellock.h (__lll_mutex_lock):
	Use atomic_exchange_and_add instead of __lll_add.
	(__lll_mutex_timedlock): Likewise.
	Patch by Ian Wienand.

2003-03-24  Steven Munroe  <sjmunroe@us.ibm.com>

	* sysdeps/unix/sysv/linux/powerpc/powerpc64/sysdep-cancel.h
	(SINGLE_THREAD_P): Fix typo.
	* tst-cancel-wrappers.sh: Handle '.'ed symbols.

2003-03-31  Ulrich Drepper  <drepper@redhat.com>

	* Makefile (tests): Add tst-align.
	* tst-align.c: New file.
	* sysdeps/i386/Makefile: Define CFLAGS-tst-align.
	* sysdeps/x86_64/Makefile: Likewise.

	* tst-tsd2.c: Add casts to avoid warnings.

+++ nptl/tst-tsd2.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2002 Free Software Foundation, Inc.
+/* Copyright (C) 2002, 2003 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
    Contributed by Ulrich Drepper <drepper@redhat.com>, 2002.
 
@@ -38,7 +38,7 @@ destr (void *arg)
 static void *
 tf (void *arg)
 {
-  pthread_key_t key = (pthread_key_t) arg;
+  pthread_key_t key = (pthread_key_t) (long int) arg;
   int err;
 
   err = pthread_setspecific (key, (void *) -2l);
@@ -65,7 +65,7 @@ do_test (void)
 
   result = 1;
 
-  err = pthread_create (&th, NULL, tf, (void *) key);
+  err = pthread_create (&th, NULL, tf, (void *) (long int) key);
   if (err != 0)
     {
       printf ("create failed: %s\n", strerror (err));
